[Frame] merge side-panel open hooks into single priority-based hook
Replaces useAutoOpenFilesPanel + useAutoOpenInteractiveContent with a single
useAutoOpenSidePanel that owns priority explicitly: interactive content (Frame)
first, file explorer second. Fixes race where the file explorer steals the
panel when a Frame is opening.
