+++ front/components/assistant/conversation/AgentMessage.tsx
@@ -12,10 +12,9 @@ import { DeletedMessage } from "@app/components/assistant/conversation/DeletedMe
 import { ErrorMessage } from "@app/components/assistant/conversation/ErrorMessage";
 import type { FeedbackSelectorBaseProps } from "@app/components/assistant/conversation/FeedbackSelector";
 import { FeedbackSelector } from "@app/components/assistant/conversation/FeedbackSelector";
-import { useAutoOpenFilesPanel } from "@app/components/assistant/conversation/files_panel/useAutoOpenFilesPanel";
 import { useGenerationContext } from "@app/components/assistant/conversation/GenerationContextProvider";
 import { getModelWithReasoningEffortLabel } from "@app/components/assistant/conversation/input_bar/modelPickerUtils";
-import { useAutoOpenInteractiveContent } from "@app/components/assistant/conversation/interactive_content/useAutoOpenInteractiveContent";
+import { useAutoOpenSidePanel } from "@app/components/assistant/conversation/useAutoOpenSidePanel";
 import type {
   AgentMessageStateWithControlEvent,
   AgentMessageWithStreaming,
@@ -1387,14 +1386,7 @@ function AgentMessageContent({
     ]
   );
 
-  // Auto-open interactive content drawer when interactive files are available.
-  const { interactiveFiles } = useAutoOpenInteractiveContent({
-    agentMessage,
-    isLastMessage,
-  });
-
-  // Auto-open file explorer panel when regular generated files are available.
-  useAutoOpenFilesPanel({ agentMessage, isLastMessage });
+  const { interactiveFiles } = useAutoOpenSidePanel({ agentMessage, isLastMessage });
 
   const blockedActionElement = blockedAction ? (
     <BlockedAction
